Добавляет изменение в сÐние от module5-task2
